+++ crates/ruff_linter/src/checkers/information_flow/mod.rs
@@ -3,3 +3,5 @@ pub(crate) mod information_flow_state;
 pub(crate) mod label;
 
 pub(crate) mod principals;
+
+pub(crate) mod helper;
