+++ crates/ruff_linter/src/rules/information_flow/rules/basic/must_include_variable_label.rs
@@ -4,13 +4,10 @@ use ruff_python_ast::{Expr, ExprName, ExprTuple};
 use ruff_source_file::Locator;
 use ruff_text_size::TextRange;
 
-use crate::{
-    checkers::{
+use crate::checkers::{
         ast::Checker,
         information_flow::{information_flow_state::read_variable_label_from_source, label::Label},
-    },
-    rules::information_flow::rules::helpers::get_variable_label_by_name,
-};
+    };
 
 /// ## What it does
 /// Check if variables have explicit labels.
