+++ crates/ruff_linter/src/rules/information_flow/rules/implicit/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) use implicit_variable_assignment::*;
+mod implicit_variable_assignment;
