+++ crates/ruff_linter/src/rules/information_flow/rules/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) use basic::*;
 pub(crate) use explicit::*;
-mod helpers;
+pub(crate) use implicit::*;
 
 mod basic;
 mod explicit;
+mod implicit;
